[RNG] Apply comments: more info in backend selector exceptions, move rng.hpp, add new line at the end of the files
